requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
djangorestframework==3.13.1
Pillow==9.1.0
psycopg2==2.9.3
python-decouple==3.6
python-dotenv==0.20.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
whitenoise==6.0.0
